learner_profile_coursegen2.o/requirements.txt
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
langchain
langchain-community
faiss-cpu
sentence-transformers
huggingface_hub
pymupdf
python-docx
requests
